feat: label handle -> (asm/transcription/parameters/label_find.c)

+++ asm/transcription/parameters/label_find.c
@@ -34,7 +34,7 @@ extern unsigned int one_label_handling(const char *param, file_t *file,
 {
     param = remove_char(param, '%');
     param = remove_char(param, ':');
-    for (int i = 0; i != file->nbLabels; i++)
+    for (size_t i = 0; i != file->nbLabels; i++)
         if (my_strcmp(param, file->label_find[i].name))  {
             value = (file->label_find[i].adresse - file->champ[k].offset + 1);
         }
